Back_Flask/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
black==24.1.1
blinker==1.7.0
boto3==1.34.44
botocore==1.34.44
Brotli==1.0.9
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
cryptography==42.0.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flake8==7.0.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
#gmpy2==2.1.2
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googletrans==4.0.0rc1
greenlet==3.0.3
grpcio==1.60.1
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==5.2.0
identify==2.5.34
idna==2.10
imageio==2.34.0
inflect==7.0.0
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.15.0
lazy_loader==0.3
Levenshtein==0.25.0
libclang==16.0.6
llvmlite==0.42.0
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nodeenv==1.8.0
numba==0.59.0
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.17.0
openai==0.28.1
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
pooch==1.8.0
pre-commit==3.6.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
Pygments==2.17.2
PyMatting==1.1.12
pymongo==4.6.1
PyMySQL==1.1.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
#pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
referencing==0.33.0
regex==2023.12.25
rembg==2.0.54
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
stack-data==0.6.3
sympy==1.12
termcolor==2.4.0
tifffile==2024.2.12
tokenizers==0.15.2
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.2
typing_extensions==4.9.0
urllib3==2.0.7
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.41.2
win-inet-pton==1.1.0
wrapt==1.14.1
yarl==1.9.4
